refactor: Address code review comments
Improvements based on code review feedback:

1. Remove sys.path modifications
   - Tests now import directly from installed package
   - Assumes package is installed in editable mode (pip install -e .)

2. Create conftest.py with shared fixtures
   - Moved simple_model, batch_data, temp_log_dir to conftest.py
   - Eliminates duplication across test files
   - Ensures consistent test fixtures

3. Fix meaningless assertions
   - Changed 'len(...) >= 0' to 'len(...) == 0'
   - Now properly verifies that sample_indices are cleared after save

4. Remove line number references
   - Replaced specific line references with behavior descriptions
   - Makes tests more robust to code changes
   - Focuses on API behavior rather than implementation details

These changes make tests more maintainable and less brittle.
