+++ tests/test_lmcleaner_core.py
@@ -13,10 +13,6 @@ import torch
 import torch.nn as nn
 from pathlib import Path
 import tempfile
-import sys
-
-# Add src to path
-sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
 
 from trainer.unlearn.lmcleaner_core import (
     AuditRecord,
@@ -40,38 +36,7 @@ from trainer.unlearn.lmcleaner_core import (
 # Test Fixtures
 # ============================================================================
 
-@pytest.fixture
-def simple_model():
-    """Create a simple linear model for testing"""
-    class SimpleModel(nn.Module):
-        def __init__(self):
-            super().__init__()
-            self.linear1 = nn.Linear(10, 5)
-            self.linear2 = nn.Linear(5, 2)
-
-        def forward(self, input_ids, labels=None, **kwargs):
-            x = self.linear1(input_ids.float())
-            logits = self.linear2(x)
-            loss = None
-            if labels is not None:
-                loss = nn.functional.cross_entropy(
-                    logits.view(-1, logits.size(-1)),
-                    labels.view(-1)
-                )
-            return type('Output', (), {'logits': logits, 'loss': loss})()
-
-    model = SimpleModel()
-    return model
-
-
-@pytest.fixture
-def batch_data():
-    """Create sample batch data"""
-    return {
-        'input_ids': torch.randn(4, 10),
-        'labels': torch.randint(0, 2, (4,))
-    }
-
+# simple_model and batch_data fixtures are defined in conftest.py
 
 @pytest.fixture
 def hvp_config():
